+++ src/view/Fillecreitrorview.java
@@ -4,59 +4,18 @@ import javax.swing.*;
 
 public class Fillecreitrorview {
     private JPanel panel1;
-    private JLabel ivlabel;
-    private JComboBox ivselektorcombox;
-    private JPanel ivselekterpanel;
-    private JPanel pasvirdpanel;
-    private JLabel Passwirde;
-    private JButton selektpasbutton;
-    private JButton geniraitepasbutton;
-    private JTextField pasvirddisplayfield;
-    private JPanel filesilektorpanel;
-    private JLabel filemesigelabel;
-    private JButton fileinbutton;
-    private JTextField fileintextfild;
-    private JCheckBox fileoutcheckBox;
-    private JButton fileoutbutton;
-    private JTextField fileouttextField;
-
-    public JPanel getPanel1() {
-        return panel1;
-    }
-
-    public JComboBox getIvselektorcombox() {
-        return ivselektorcombox;
-    }
-
-    public JButton getSelektpasbutton() {
-        return selektpasbutton;
-    }
-
-    public JButton getGeniraitepasbutton() {
-        return geniraitepasbutton;
-    }
-
-    public JTextField getPasvirddisplayfield() {
-        return pasvirddisplayfield;
-    }
-
-    public JButton getFileinbutton() {
-        return fileinbutton;
-    }
-
-    public JTextField getFileintextfild() {
-        return fileintextfild;
-    }
-
-    public JCheckBox getFileoutcheckBox() {
-        return fileoutcheckBox;
-    }
-
-    public JButton getFileoutbutton() {
-        return fileoutbutton;
-    }
-
-    public JTextField getFileouttextField() {
-        return fileouttextField;
-    }
+    private JPanel filegenirator;
+    private JPanel siseinput;
+    private JPanel creationpanel;
+    private JTextField DirektorytextField;
+    private JTextField FileNamntextField;
+    private JButton DirectoryButton;
+    private JButton TestFilebutton;
+    private JLabel DirectoryStatesLabel;
+    private JLabel FileNameTestStateslabel;
+    private JLabel GBlabel;
+    private JTextField GBtextField;
+    private JButton Okbutton;
+    private JLabel StatusLabel;
+    private JLabel StatusValueLabel;
 }
